+++ shuup/admin/modules/products/forms/package_forms.py
@@ -10,18 +10,32 @@ from __future__ import unicode_literals
 import six
 from django import forms
 from django.contrib import messages
+from django.db.models import ObjectDoesNotExist
 from django.db.transaction import atomic
 from django.utils.translation import ugettext_lazy as _
 
 from shuup.admin.forms.widgets import PackageProductChoiceWidget
 from shuup.admin.modules.products.utils import clear_existing_package
 from shuup.core.excs import ImpossibleProductModeException, Problem
-from shuup.core.models import Product
+from shuup.core.models import Product, Shop
 
 from .parent_forms import ProductChildBaseFormSet
 
 
 class PackageChildForm(forms.Form):
+    def __init__(self, **kwargs):
+        initial = kwargs.get("initial", {})
+        self.product = initial.get("child")
+        if self.product:
+            self.shop_products = []
+            for shop in Shop.objects.all():
+                try:
+                    shop_product = self.product.get_shop_instance(shop)
+                    self.shop_products.append(shop_product)
+                except ObjectDoesNotExist:
+                    continue
+        super(PackageChildForm, self).__init__(**kwargs)
+
     child = forms.ModelChoiceField(
         queryset=Product.objects.all(),
         widget=PackageProductChoiceWidget(),
@@ -34,6 +48,21 @@ class PackageChildForm(forms.Form):
         widget=forms.NumberInput(attrs={"class": "form-control"}),
         required=True,
     )
+
+    def get_stock_statuses(self):
+        stocks = {}
+        if not self.product:
+            return stocks
+        sales_unit = self.product.sales_unit
+        sales_decimals = sales_unit.decimals if sales_unit else 0
+        sales_unit_short_name = sales_unit.short_name if sales_unit else ""
+        for shop_product in self.shop_products:
+            for supplier in shop_product.suppliers.all():
+                if supplier in stocks.keys():
+                    continue
+                stock_status = supplier.get_stock_status(product_id=self.product.id)
+                stocks[supplier] = (supplier, stock_status, sales_decimals, sales_unit_short_name)
+        return stocks
 
 
 class PackageChildFormSet(ProductChildBaseFormSet):